requirements.txt
python-telegram-bot==20.8
APScheduler
pandas
matplotlib
requests